Add Flatten layer

+++ crates/nove_model/src/nn/gru.rs
@@ -519,7 +519,7 @@ impl Model for Gru {
 
                 for t in sequence_indices {
                     // Get input for this time step
-                    let input_slice = layer_input.narrow(0, t, 1)?.squeeze(Some(0))?;
+                    let input_slice = layer_input.narrow(0, t as isize, 1)?.squeeze(Some(0))?;
 
                     // Forward through GRU cell
                     current_hidden_state = cell.forward((input_slice, current_hidden_state))?;
